Did some renaming so that BOSS and SP are the latest versions and other versions are moved out of the way and disabled.
