+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/algorithm/oracle/cpdag/SP.java
@@ -1,32 +1,31 @@
 package edu.cmu.tetrad.algcomparison.algorithm.oracle.cpdag;
 
 import edu.cmu.tetrad.algcomparison.algorithm.Algorithm;
-import edu.cmu.tetrad.algcomparison.independence.IndependenceWrapper;
 import edu.cmu.tetrad.algcomparison.score.ScoreWrapper;
 import edu.cmu.tetrad.algcomparison.utils.HasKnowledge;
-import edu.cmu.tetrad.algcomparison.utils.TakesIndependenceWrapper;
 import edu.cmu.tetrad.algcomparison.utils.UsesScoreWrapper;
 import edu.cmu.tetrad.annotation.AlgType;
 import edu.cmu.tetrad.annotation.Bootstrapping;
+import edu.cmu.tetrad.annotation.Experimental;
 import edu.cmu.tetrad.data.DataModel;
-import edu.cmu.tetrad.data.DataSet;
 import edu.cmu.tetrad.data.DataType;
 import edu.cmu.tetrad.data.Knowledge;
 import edu.cmu.tetrad.graph.EdgeListGraph;
 import edu.cmu.tetrad.graph.Graph;
-import edu.cmu.tetrad.search.IndependenceTest;
+import edu.cmu.tetrad.search.PermutationSearch2;
 import edu.cmu.tetrad.search.Score;
+import edu.cmu.tetrad.search.Sp;
 import edu.cmu.tetrad.util.Parameters;
 import edu.cmu.tetrad.util.Params;
-import edu.pitt.dbmi.algo.resampling.GeneralResamplingTest;
 
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * BOSS (Best Order Score Search).
+ * SP (Sparsest Permutation)
  *
- * @author jdramsey
+ * @author bryanandrews
+ * @author josephramsey
  */
 @edu.cmu.tetrad.annotation.Algorithm(
         name = "SP",
@@ -34,45 +33,37 @@ import java.util.List;
         algoType = AlgType.forbid_latent_common_causes
 )
 @Bootstrapping
-public class SP implements Algorithm, UsesScoreWrapper, TakesIndependenceWrapper, HasKnowledge {
+@Experimental
+public class SP implements Algorithm, UsesScoreWrapper, HasKnowledge {
     static final long serialVersionUID = 23L;
-    private ScoreWrapper score = null;
-    private IndependenceWrapper test;
-
+    private ScoreWrapper score;
     private Knowledge knowledge = new Knowledge();
 
     public SP() {
         // Used in reflection; do not delete.
     }
 
-    public SP(IndependenceWrapper test, ScoreWrapper score) {
-        this.test = test;
+    public SP(ScoreWrapper score) {
         this.score = score;
     }
 
+
     @Override
     public Graph search(DataModel dataModel, Parameters parameters) {
-        if (parameters.getInt(Params.NUMBER_RESAMPLING) < 1) {
-            Score score = this.score.getScore(dataModel, parameters);
+        Score score = this.score.getScore(dataModel, parameters);
+        PermutationSearch2 permutationSearch2 = new PermutationSearch2(new Sp(score));
 
-            IndependenceTest test = this.test.getTest(dataModel, parameters);
+//        int numPerTier = 10;
+//        for (int tier = 0; tier < 2; tier++) {
+//            for (int i = 1; i <= numPerTier; i++) {
+//                this.knowledge.addToTier(tier, "X" + (numPerTier * tier + i));
+//            }
+//        }
 
-            test.setVerbose(parameters.getBoolean(Params.VERBOSE));
+        permutationSearch2.setKnowledge(this.knowledge);
+        permutationSearch2.setVerbose(parameters.getBoolean(Params.VERBOSE));
 
-            edu.cmu.tetrad.search.SP sp = new edu.cmu.tetrad.search.SP(test, score);
-            sp.setKnowledge(knowledge);
-            sp.bestOrder(score.getVariables());
-            return sp.getGraph(true);
-        } else {
-            SP algorithm = new SP(this.test, this.score);
-
-            DataSet data = (DataSet) dataModel;
-            GeneralResamplingTest search = new GeneralResamplingTest(data, algorithm, parameters.getInt(Params.NUMBER_RESAMPLING), parameters.getDouble(Params.PERCENT_RESAMPLE_SIZE), parameters.getBoolean(Params.RESAMPLING_WITH_REPLACEMENT), parameters.getInt(Params.RESAMPLING_ENSEMBLE), parameters.getBoolean(Params.ADD_ORIGINAL_DATASET));
-
-            search.setParameters(parameters);
-            search.setVerbose(parameters.getBoolean(Params.VERBOSE));
-            return search.search();
-        }
+        return permutationSearch2.search();
     }
 
     @Override
@@ -82,40 +73,33 @@ public class SP implements Algorithm, UsesScoreWrapper, TakesIndependenceWrapper
 
     @Override
     public String getDescription() {
-        return "SP (Sparsest Permutation) using " + test.getDescription() + " or " + score.getDescription();
+        return "SP (Sparsest Permutation) using " + this.score.getDescription();
     }
 
     @Override
     public DataType getDataType() {
-        return score.getDataType();
+        return this.score.getDataType();
     }
 
     @Override
     public List<String> getParameters() {
         ArrayList<String> params = new ArrayList<>();
-        params.add(Params.GRASP_USE_RASKUTTI_UHLER);
+
+        // Flags
         params.add(Params.VERBOSE);
+
+
         return params;
     }
 
     @Override
     public ScoreWrapper getScoreWrapper() {
-        return score;
+        return this.score;
     }
 
     @Override
     public void setScoreWrapper(ScoreWrapper score) {
         this.score = score;
-    }
-
-    @Override
-    public IndependenceWrapper getIndependenceWrapper() {
-        return test;
-    }
-
-    @Override
-    public void setIndependenceWrapper(IndependenceWrapper independenceWrapper) {
-        this.test = independenceWrapper;
     }
 
     @Override
@@ -125,7 +109,6 @@ public class SP implements Algorithm, UsesScoreWrapper, TakesIndependenceWrapper
 
     @Override
     public void setKnowledge(Knowledge knowledge) {
-        this.knowledge = new Knowledge((Knowledge) knowledge);
+        this.knowledge = knowledge;
     }
-
 }
