+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/algorithm/oracle/cpdag/SP_OLD.java
@@ -1,0 +1,130 @@
+package edu.cmu.tetrad.algcomparison.algorithm.oracle.cpdag;
+
+import edu.cmu.tetrad.algcomparison.algorithm.Algorithm;
+import edu.cmu.tetrad.algcomparison.independence.IndependenceWrapper;
+import edu.cmu.tetrad.algcomparison.score.ScoreWrapper;
+import edu.cmu.tetrad.algcomparison.utils.HasKnowledge;
+import edu.cmu.tetrad.algcomparison.utils.TakesIndependenceWrapper;
+import edu.cmu.tetrad.algcomparison.utils.UsesScoreWrapper;
+import edu.cmu.tetrad.annotation.Bootstrapping;
+import edu.cmu.tetrad.data.DataModel;
+import edu.cmu.tetrad.data.DataSet;
+import edu.cmu.tetrad.data.DataType;
+import edu.cmu.tetrad.data.Knowledge;
+import edu.cmu.tetrad.graph.EdgeListGraph;
+import edu.cmu.tetrad.graph.Graph;
+import edu.cmu.tetrad.search.IndependenceTest;
+import edu.cmu.tetrad.search.Score;
+import edu.cmu.tetrad.util.Parameters;
+import edu.cmu.tetrad.util.Params;
+import edu.pitt.dbmi.algo.resampling.GeneralResamplingTest;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * BOSS (Best Order Score Search).
+ *
+ * @author jdramsey
+ */
+//@edu.cmu.tetrad.annotation.Algorithm(
+//        name = "SP-OLD",
+//        command = "sp-old",
+//        algoType = AlgType.forbid_latent_common_causes
+//)
+@Bootstrapping
+public class SP_OLD implements Algorithm, UsesScoreWrapper, TakesIndependenceWrapper, HasKnowledge {
+    static final long serialVersionUID = 23L;
+    private ScoreWrapper score = null;
+    private IndependenceWrapper test;
+
+    private Knowledge knowledge = new Knowledge();
+
+    public SP_OLD() {
+        // Used in reflection; do not delete.
+    }
+
+    public SP_OLD(IndependenceWrapper test, ScoreWrapper score) {
+        this.test = test;
+        this.score = score;
+    }
+
+    @Override
+    public Graph search(DataModel dataModel, Parameters parameters) {
+        if (parameters.getInt(Params.NUMBER_RESAMPLING) < 1) {
+            Score score = this.score.getScore(dataModel, parameters);
+
+            IndependenceTest test = this.test.getTest(dataModel, parameters);
+
+            test.setVerbose(parameters.getBoolean(Params.VERBOSE));
+
+            edu.cmu.tetrad.search.SP_Old sp = new edu.cmu.tetrad.search.SP_Old(test, score);
+            sp.setKnowledge(knowledge);
+            sp.bestOrder(score.getVariables());
+            return sp.getGraph(true);
+        } else {
+            SP_OLD algorithm = new SP_OLD(this.test, this.score);
+
+            DataSet data = (DataSet) dataModel;
+            GeneralResamplingTest search = new GeneralResamplingTest(data, algorithm, parameters.getInt(Params.NUMBER_RESAMPLING), parameters.getDouble(Params.PERCENT_RESAMPLE_SIZE), parameters.getBoolean(Params.RESAMPLING_WITH_REPLACEMENT), parameters.getInt(Params.RESAMPLING_ENSEMBLE), parameters.getBoolean(Params.ADD_ORIGINAL_DATASET));
+
+            search.setParameters(parameters);
+            search.setVerbose(parameters.getBoolean(Params.VERBOSE));
+            return search.search();
+        }
+    }
+
+    @Override
+    public Graph getComparisonGraph(Graph graph) {
+        return new EdgeListGraph(graph);
+    }
+
+    @Override
+    public String getDescription() {
+        return "SP-OLD (Sparsest Permutation) using " + test.getDescription() + " or " + score.getDescription();
+    }
+
+    @Override
+    public DataType getDataType() {
+        return score.getDataType();
+    }
+
+    @Override
+    public List<String> getParameters() {
+        ArrayList<String> params = new ArrayList<>();
+        params.add(Params.GRASP_USE_RASKUTTI_UHLER);
+        params.add(Params.VERBOSE);
+        return params;
+    }
+
+    @Override
+    public ScoreWrapper getScoreWrapper() {
+        return score;
+    }
+
+    @Override
+    public void setScoreWrapper(ScoreWrapper score) {
+        this.score = score;
+    }
+
+    @Override
+    public IndependenceWrapper getIndependenceWrapper() {
+        return test;
+    }
+
+    @Override
+    public void setIndependenceWrapper(IndependenceWrapper independenceWrapper) {
+        this.test = independenceWrapper;
+    }
+
+    @Override
+    public Knowledge getKnowledge() {
+        return this.knowledge;
+    }
+
+    @Override
+    public void setKnowledge(Knowledge knowledge) {
+        this.knowledge = new Knowledge((Knowledge) knowledge);
+    }
+
+}
